change(nn): split worker code into separate non-worker file, which is exported by the worker
